Create Dish.py
*Note: Incomplete*
The code helps build the back-end system that processes data related to a possible menu for a restaurant. I hardcoded a few dishes to go the menu and included data regarding the ingredients, labor, difficulty, prices for ingredients, prices for the dish, etc. This data can be used to find profits for a dish. Benefits include: speeding up accounting by being able to change one thing and not needing complete recalculation by an individual (program can do it). 
Goals to include in this project:
1. Get the connection to an excel file to work properly so all data can be from a spreadsheet and not hardcoded
2.